+++ constcheck/_typing.py
@@ -16,4 +16,6 @@ PathFileStringRep = _t.Dict[_Path, FileStringRep]
 PathLike = _t.Union[str, _os.PathLike]
 TokenList = _t.List[_TokenText]
 ValueTuple = _t.Tuple[int, int]
-ArgTuple = _t.Tuple[PathLike, ValueTuple, bool, bool, _t.Optional[str]]
+ArgTuple = _t.Tuple[
+    PathLike, ValueTuple, bool, bool, _t.Optional[str], _t.List[str]
+]